feat: MVVM Architecture
